scripts/env-collect: Poll wttr.in less often
Change to the script to make it more friendly for conference
demos where there may be a lot of RF noise. The wttr.in site is
polled once per minute until a valid response. After that it is
only polled for an update once per hour. If that update is missed
then continue to use old data rather than show an error. In order
to reduce stale data, attempt an update of the weather each minute
loop if the hour update is missed but continue to show old data
rather than an error.

The rest of the interface is still updated once per minute